Original Loops File
Grade: 77/100
Comments: All of you javadoc comments were inside of the method, and therefore not javadocs. It should be like this: /** * Calculates _______ by doing _____ * @param var - an integer used in the method for ______ * @return the count of _______ (or something like that) */ public int methodName (int var) {...} Overall it seems like you did ok, but got stuck on the last two methods especially. When that happens, please come to help hours! And if you feel like you're getting behind on conceptual stuff, feel free to sign up for a one on one help session with a TA.